refactor: Change the directory organization.
I have been working with this codes and I notice it is important to separate the base codes or modules from complete projects because they are more complex and also use the module codes.his codes are useful to build complete projects and they work as small pieces that accomplish basic functions,

So the current repository will keep only changes on this modules and new developments that will help to build future projects.

The Cayenne project on MITA has been send to a separate repository to keep tracking of improvements and it has a specific purporse, sending complete data using WiFi and MQTT protovol to the IoT platform Cayenne.